# swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ova-journey-api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "JourneyApiV1"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/journeys": {
      "get": {
        "operationId": "JourneyApiV1_ListJourneysV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiListJourneysResponseV1"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "JourneyApiV1"
        ]
      },
      "post": {
        "operationId": "JourneyApiV1_CreateJourneyV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiCreateJourneyResponseV1"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiCreateJourneyRequestV1"
            }
          }
        ],
        "tags": [
          "JourneyApiV1"
        ]
      }
    },
    "/v1/journeys/multi": {
      "post": {
        "operationId": "JourneyApiV1_MultiCreateJourneyV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiMultiCreateJourneyResponseV1"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiMultiCreateJourneyRequestV1"
            }
          }
        ],
        "tags": [
          "JourneyApiV1"
        ]
      }
    },
    "/v1/journeys/{journeyId}": {
      "get": {
        "operationId": "JourneyApiV1_DescribeJourneyV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiDescribeJourneyResponseV1"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "JourneyApiV1"
        ]
      },
      "delete": {
        "operationId": "JourneyApiV1_RemoveJourneyV1",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "journeyId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "JourneyApiV1"
        ]
      }
    }
  },
  "definitions": {
    "apiCreateJourneyRequestV1": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "uint64"
        },
        "address": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "startTime": {
          "type": "string",
          "format": "date-time"
        },
        "endTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiCreateJourneyResponseV1": {
      "type": "object",
      "properties": {
        "journeyId": {
          "type": "string",
          "format": "uint64"
        }
      }
    },
    "apiDescribeJourneyResponseV1": {
      "type": "object",
      "properties": {
        "journey": {
          "$ref": "#/definitions/apiJourney"
        }
      }
    },
    "apiJourney": {
      "type": "object",
      "properties": {
        "journeyId": {
          "type": "string",
          "format": "uint64"
        },
        "userId": {
          "type": "string",
          "format": "uint64"
        },
        "address": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "startTime": {
          "type": "string",
          "format": "date-time"
        },
        "endTime": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "apiListJourneysResponseV1": {
      "type": "object",
      "properties": {
        "journeys": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiJourney"
          }
        }
      }
    },
    "apiMultiCreateJourneyRequestV1": {
      "type": "object",
      "properties": {
        "journeys": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apiCreateJourneyRequestV1"
          }
        }
      }
    },
    "apiMultiCreateJourneyResponseV1": {
      "type": "object",
      "properties": {
        "journeyIds": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "uint64"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "typeUrl": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}